+++ src/main/java/lotto/model/generator/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.generator;
+package lotto.model.generator;
 
 import java.util.ArrayList;
 import java.util.List;
