+++ src/main/java/lotto/model/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.generator;
+package lotto.model.generator;
 
 public interface NumberGenerator {
     int generate(int maxNumber);
